Convert WorkerMessage to an enum. Replace worker_id usize with u32
WorkerMessage had a bunch of logic to determine the kind of message when
we can just use an enum.

worker_uid was a usize, this doesn't make sense because a usize is for
indexing a slice, worker_uid is just an identifier so u32 is fine

+++ src/config.rs
@@ -5,11 +5,11 @@ use argparse::{ArgumentParser, Store, StoreFalse, StoreOption, StoreTrue};
 use directories::BaseDirs;
 use fancy_regex::Regex;
 use serde::{Deserialize, Serialize};
-use tracing::debug;
 use std::fmt;
 use std::fs;
 use std::path::Path;
 use std::path::PathBuf;
+use tracing::debug;
 use tracing::error;
 
 ///This struct contains any system specific data (paths, extensions, etc)
@@ -29,7 +29,7 @@ pub struct ServerConfig {
 pub struct WorkerConfig {
     pub server_address: String,
     pub server_port: u16,
-    pub uid: Option<usize>,
+    pub uid: Option<u32>,
     config_path: PathBuf,
 }
 
@@ -78,7 +78,7 @@ impl WorkerConfig {
         config
     }
 
-    pub fn insert_uid(&mut self, uid: usize) {
+    pub fn insert_uid(&mut self, uid: u32) {
         self.uid = Some(uid);
     }
